refactor: shell panel, add prettier config
