+++ electron/src/frontend/components/agent-workspace/agent-header.tsx
@@ -1,12 +1,7 @@
 'use client'
 import { useState, useEffect } from 'react'
-import {
-    GitPullRequest,
-    LayoutPanelLeft,
-    Columns2,
-} from 'lucide-react'
+import { GitPullRequest, LayoutPanelLeft, Columns2 } from 'lucide-react'
 import { Switch } from '@/components/ui/switch'
-import { ViewMode } from '@/lib/types'
 
 const AgentWorkspaceHeader = ({
     viewMode,
@@ -78,7 +73,9 @@ export const ToggleTimelineHeader = ({
     return (
         <div className="flex flex-row gap-2 items-center mr-5 absolute right-0">
             <button
-                className={`border border-neutral-500 rounded-md pl-4 pr-3 flex p-2 items-center justify-center rounded-md transition duration-200 hover:bg-gray-100 dark:hover:bg-batman ${showTimeline ? 'bg-gray-100 dark:bg-batman' : ''}`}
+                className={`border border-neutral-500 rounded-md pl-4 pr-3 flex p-2 items-center justify-center rounded-md transition duration-200 hover:bg-gray-100 dark:hover:bg-batman ${
+                    showTimeline ? 'bg-gray-100 dark:bg-batman' : ''
+                }`}
                 onClick={() => setShowTimeline(!showTimeline)}
             >
                 <p className="mr-2 font-bold">Timeline</p>
